Remove NPM's package-lock.json from Git.
There are some pros/cons to having this file. See https://github.com/npm/npm/issues/20603

In practice, when I make a dependency change in package.json and look at the diff, I see 100's of differences in this file making it hard to see what else is changing. At the same time, I don't want to commit it before it's final.

If this becomes an issue later on, we can re-add it.