Use properties the correct way. Update instructions to set openai.api.key.
The way that we were using property files was incorrect. This commit adds the necessary hooks to have the Spring framework read these files for us.

It verifies that the new way is working by using an openai.api.key property specified in an optional ${HOME}/project_leo.properties file. The OpenAI test passes whether the property is set or not. When not set, the test is skipped.