Add teacher and student tables.
I chose to have a foreign key in users point to a student and teacher in order to allow a load of users to indicate what permissions someone has. If the foreign keys are null, then the user doesn't have that permission.

This does allow for the possibility of dangling students and teachers, which is undesirable. But, that can be removed when scrubbing.